Map place id to station UIC
Trip and Places model documentation states that `using the UIC code set is highly recommended and mandatory for train stations`:

```
StopPlace: represent a place where a train or a bus stops. It is obviously the most relevant type for OSDM. StopPlaces can be indicated as codes from different code sets. As with other code list based representations in the API, using the UIC code set is highly recommended and mandatory for train stations.
```

https://osdm.io/spec/models/#TripsAndPlaces